pre final

+++ CODIGO/PROVEEDORES/deleteProveedores.py
@@ -12,15 +12,19 @@ absolutepath = os.path.abspath(__file__)
 
 from CODIGO.BD import queryFunctions
 
-def eliminarProveedor(proveedor):
-    queryFunctions.updateBD('DELETE FROM proveedores WHERE idproovedor = %s' % proveedor)
+def eliminarProveedor(proveedor, nombreProveedor):
+    salidaComprobacionProductos = queryFunctions.selectBD('SELECT * FROM PRODUCTOS WHERE proveedorCif = (SELECT cif FROM proveedores where nombre = "%s")' % nombreProveedor)
+    
+    if len(salidaComprobacionProductos) > 0:
+        sg.Popup('Provider can not be delete while has referenced products')
+    else:
+        queryFunctions.updateBD('DELETE FROM proveedores WHERE idproovedor = %s' % proveedor)
 
 def getValores():
     return queryFunctions.selectBD('SELECT idproovedor as ID, nombre as NAME, cif as CIF, telefono as PHONE, email as EMAIL, direccion as ADDRESS, cuentabanco as ACCOUNT FROM proveedores')
 
 
 valores = queryFunctions.selectBD('SELECT idproovedor as ID, nombre as NAME, cif as CIF, telefono as PHONE, email as EMAIL, direccion as ADDRESS, cuentabanco as ACCOUNT FROM proveedores')
-print(valores.to_dict())
 
 sg.theme('DarkGrey6')
 
@@ -30,7 +34,7 @@ layout = [
     [sg.Button('BACK TO MENU', key='backToMenu', size=(20,1))]
 ]
 
-window = sg.Window('DELETE PROVIDER', layout, size=(800,280), element_justification='c')
+window = sg.Window('DELETE PROVIDER', layout, size=(800,280), element_justification='c',icon=os.path.join(absolutepath, '..\\..\\..\\RESOURCES\\AppIcon\\icon.ico'))
 
 while True:             
     event, values = window.read()
@@ -40,19 +44,19 @@ while True:
         event()
     elif event == 'deleteOne':
         try:
-            eliminarProveedor(getValores().values.tolist()[values['table'][0]][0])
-            window['table'].update(getValores().values.tolist())
+            if len(values['table']) != 0:
+                eliminarProveedor(getValores().values.tolist()[values['table'][0]][0], getValores().values.tolist()[values['table'][0]][1])
+                window['table'].update(getValores().values.tolist())
+            else:
+                sg.Popup('You must select one provider on the table')
         except Exception as ex:
             print(ex)
     elif event == 'deleteAll':
-        print('Eliminar todos')
         salida = sg.popup_yes_no('Do you want to delete all? This action can not be undone', title='Delete All')
-        print(salida)
         if salida == 'Yes':
             queryFunctions.updateBD('DELETE FROM PROVEEDORES')
             window['table'].update(getValores().values.tolist())
 
     elif event == 'backToMenu':
-        print('Back to menu')
         window.close()
         subprocess.call(['python', os.path.join(absolutepath, '..\\..\\MENUS\\menuProveedores.py')])
